Revert to newest state after fetching old state
Whenever a plot is made of a hypervisor and the improvement it makes,
the hypervisor state is reverted back to its initial state, so the state
after migrations is lost.

This commit makes sure the hypervisor state is not changed while
generating a plot by reverting the hypervisor to its newest state.